scraper/requirements.txt
# Core Scraping
curl-cffi==0.6.2
beautifulsoup4==4.12.3
lxml==5.3.0

# Playwright (for Tier 2/3)
playwright==1.48.0
playwright-stealth==1.0.5

# Apify Integration
apify-client==1.7.1

# Data Processing
pandas==2.2.3
pydantic==2.9.2
python-dateutil==2.9.0

# Orchestration
prefect==3.1.9
redis==5.2.0
psycopg[binary]==3.2.3
asyncpg==0.30.0

# Reliability
tenacity==9.0.0
ratelimit==2.2.1

# Utilities
python-dotenv==1.0.1
loguru==0.7.3
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
